v1

+++ 038_KonuYorumCoreIntroBilgeAdam/DataAccess/Yorum.cs
@@ -8,7 +8,7 @@ namespace _038_KonuYorumCoreIntroBilgeAdam.DataAccess
         public int Id { get; set; }
         public string Icerik { get; set; }
         public string Yorumcu { get; set; }
-        public byte? Puan { get; set; }
+        public int? Puan { get; set; }
         public int KonuId { get; set; }
 
         public virtual Konu Konu { get; set; }
